♻️ refactor: Update library copyrights, update dependencies, and add .gitignore file
Updated all copyrights for the project libraries. Also included in the diff are changes to the dependency list in build.sbt where library versions were updated to the latest available ones to ensure functionalities are up to date. Inserted a .gitignore file as well to improve codebase cleanliness by avoiding unnecessary file tracking.